requirements.txt
absl-py==0.2.2
astor==0.6.2
attrs==17.4.0
bleach==1.5.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
Cython==0.28.5
enum34==1.1.6
future==0.16.0
gast==0.2.0
glfw==1.7.0
grpcio==1.12.1
gym==0.10.5
html5lib==0.9999999
idna==2.6
imageio==2.3.0
kiwisolver==1.0.1
Markdown==2.6.9
matplotlib==2.1.0
mpi4py==3.0.0
mujoco-py==1.50.1.56
numpy==1.13.3
pandas==0.21.0
Pillow==5.2.0
pluggy==0.6.0
protobuf==3.4.0
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
pyglet==1.3.2
pyparsing==2.2.0
pytest==3.4.2
python-dateutil==2.6.1
python-dotenv==0.9.1
pytz==2017.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.12.2
